travis test whether compilation works on mac and linux
This checks both gcc and clang on linux, but only clang on mac

Becuse there are so many debug warnings they are supressed in a very hacky way:
Make is run first with error output ignored.
Then an incremental build is run which has the error output, but will only build the files that errored the first time (and will error again).
If anyone knows a better way to ignore the warnings that would be great.