compile fixes and improvements
mac was failing to generate an archive.

Icons now in an app asset so it can be properly modified in the project instead of being a static .icns file.

Info.plist now gets its bundle name from the product bundle identifier field, this means it can be changed through the gui instead of having to manually modify the bundle identifier. All other settings are left blank.

Info.plist will need further updates to take values set in Xcode to make it easier to distribute apps.

OpenAL and SDL have to skip installation in release mode builds for the archive to be created correctly.