Use SDL2 Backend with OpenAL
This change implements openal to use the sdl2 backend.

The main change in this is in openal-soft cmakeLists.txt it was not using our custom finders to find sdl2 for us before and so it had to be changed to a list append instead of a set. This needs to be remembered for future updates to openal