ground work for openal refactor
since vcpkg sets up openal-soft for us we no longer have to set certain defines that control whether certain features are okay per platform. We just check ALC_EXT_EFX and load the functions as needed. More ground work to be added to this and since vcpkg is going to be rolled in this is the most logical place for it